requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
einops==0.2.0
filelock==3.0.12
gdown==3.11.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
lmdb==1.0.0
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
prefetch-generator==1.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.7.0
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.1
tifffile==2021.6.14
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
vietocr==0.3.5
